Reformat and Refactor #2
