+++ Drawing.CenterView/UI/IconHandler.cs
