+++ Drawing.CenterView/UI/UI.cs
@@ -87,7 +87,7 @@ public partial class PluginForm
 
             version.Text = @"v" + Assembly.GetEntryAssembly()?.GetName().Version.ToString();
 
-            if (_drawingHandler.GetActiveDrawing().Title3.Equals("X")) excludeCheckBox.Checked = true;
+            if (DrawingHandler.GetActiveDrawing().Title3.Equals("X")) excludeCheckBox.Checked = true;
 
             _rightArrow = new SvgIcon(SvgIcon.Icon.RightArrow);
             _rightArrow.ChangeIconColors(Color.Black, targetColor);
@@ -170,7 +170,7 @@ public partial class PluginForm
         view.GetStringUserProperties(out Dictionary<string, string> viewType);
         GetViewTypeEnum(viewType);
         InfoBox.OnInfo(infoBox, $"Shifting Right => {(ViewType)GetViewTypeEnum(viewType)}");
-        _drawingHandler.GetActiveDrawing().CommitChanges("Shift View Right");
+        DrawingHandler.GetActiveDrawing().CommitChanges("Shift View Right");
     }
 
     private void ShiftViewUp(int amount)
@@ -182,7 +182,7 @@ public partial class PluginForm
         view.GetStringUserProperties(out Dictionary<string, string> viewType);
         GetViewTypeEnum(viewType);
         InfoBox.OnInfo(infoBox, $"{(ViewType)GetViewTypeEnum(viewType)}\nShifting Up =^");
-        _drawingHandler.GetActiveDrawing().CommitChanges("Shift View Up");
+        DrawingHandler.GetActiveDrawing().CommitChanges("Shift View Up");
     }
 
     private void ShiftViewDown(int amount)
@@ -194,7 +194,7 @@ public partial class PluginForm
         view.GetStringUserProperties(out Dictionary<string, string> viewType);
         GetViewTypeEnum(viewType);
         InfoBox.OnInfo(infoBox, $"Shifting Down=v\n{(ViewType)GetViewTypeEnum(viewType)}");
-        _drawingHandler.GetActiveDrawing().CommitChanges("Shift View Down");
+        DrawingHandler.GetActiveDrawing().CommitChanges("Shift View Down");
     }
 
     private void ShiftViewLeft(int amount)
@@ -206,7 +206,7 @@ public partial class PluginForm
         view.GetStringUserProperties(out Dictionary<string, string> viewType);
         GetViewTypeEnum(viewType);
         InfoBox.OnInfo(infoBox, $"{(ViewType)GetViewTypeEnum(viewType)} <= Shifting Left");
-        _drawingHandler.GetActiveDrawing().CommitChanges("Shift View Left");
+        DrawingHandler.GetActiveDrawing().CommitChanges("Shift View Left");
     }
 }
